Commit - work 19092019
PMD
SpootBugs

+++ spring-parents/src/test/java/com/everis/parents/app/controller/ParentRestControllerTest.java
@@ -16,16 +16,30 @@ import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import reactor.core.publisher.Mono;
 
+/**
+ *. 
+ * @author lriveras
+ *
+ */
 @AutoConfigureWebTestClient
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class ParentRestControllerTest {
+
+  /**
+ * Injection WebTestClient.
+ */
   @Autowired
   private WebTestClient client;
-
+  /**
+ * Injection ParenRepository.
+ */
   @Autowired
   private ParenRepository service;
 
+  /**
+ * Test findAll.
+ */
   @Test
   public void findAll() {
     client.get().uri("/api/v1.0/parents")
@@ -34,7 +48,7 @@ public class ParentRestControllerTest {
          .expectHeader().contentType(MediaType.APPLICATION_JSON_UTF8)
          .expectBodyList(Parennt.class)
          .consumeWith(response -> {
-           List<Parennt> parents = response.getResponseBody();
+           final List<Parennt> parents = response.getResponseBody();
            parents.forEach(s -> {
              System.out.println(s.getFullname() + " - " + s.getNumberID());
            });
@@ -42,43 +56,54 @@ public class ParentRestControllerTest {
          });
   }
 
+  /**
+ * Test findByID.
+ */
   @Test
-public void show() {
-    Parennt parent = service.obtenerPorName("David").block();
-
-    client.get().uri("/api/v1.0/parents/{id}", Collections.singletonMap("id", parent.getId()))
-        .accept(MediaType.APPLICATION_JSON_UTF8)
-        .exchange()
-        .expectStatus().isOk()
-        .expectHeader()
-        .contentType(MediaType.APPLICATION_JSON_UTF8)
-        .expectBody(Parennt.class).consumeWith(response -> {
-          Parennt pr = response.getResponseBody();
-          Assertions.assertThat(pr.getId()).isNotEmpty();
-          Assertions.assertThat(pr.getId().length() > 0).isTrue();
-          Assertions.assertThat(pr.getFullname()).isEqualTo("David");
-        });
+public void findById() {
+    final Parennt parent = service.findByFullname_par("David").block();
+    if (parent != null) {
+      client.get().uri("/api/v1.0/parents/{id}", Collections.singletonMap("id", parent.getId()))
+          .accept(MediaType.APPLICATION_JSON_UTF8)
+          .exchange()
+          .expectStatus().isOk()
+          .expectHeader()
+          .contentType(MediaType.APPLICATION_JSON_UTF8)
+          .expectBody(Parennt.class).consumeWith(response -> {
+            final Parennt pr = response.getResponseBody();
+            Assertions.assertThat(pr.getId()).isNotEmpty();
+            Assertions.assertThat(pr.getId().length() > 0).isTrue();
+            Assertions.assertThat(pr.getFullname()).isEqualTo("David");
+          });
+    }
   }
 
+  /**
+ * Test save.
+ */
   @Test
-  public void crear() {
-    Parennt parennt = new Parennt("Romina", "Femenino", new Date(), "DNI", "00000000", "23");
+  public void save() {
+    final Parennt parennt = new Parennt("Rominax", "Femenino", new Date(), "DNI", "00000001", "23");
     client.post().uri("/api/v1.0/parents").contentType(MediaType.APPLICATION_JSON_UTF8)
         .accept(MediaType.APPLICATION_JSON_UTF8).body(Mono.just(parennt), Parennt.class).exchange()
         .expectStatus().isOk().expectHeader().contentType(MediaType.APPLICATION_JSON_UTF8)
         .expectBody(Parennt.class).consumeWith(response -> {
-          Parennt ps = response.getResponseBody();
-          Assertions.assertThat(ps.getId()).isNotEmpty();
-          Assertions.assertThat(ps.getFullname()).isEqualTo("Romina");
+          final Parennt paren = response.getResponseBody();
+          Assertions.assertThat(paren.getId()).isNotEmpty();
+          Assertions.assertThat(paren.getFullname()).isEqualTo("Rominax");
         });
   }
 
+  /**
+ * Test update.
+ */
   @Test
-  public void editarTest() {
-    Parennt parent = service.obtenerPorName("Sandrox").block();
-    Parennt parentEditado = new Parennt("Sandro", "Masculino", new Date(),"DNI", "20020711","22");
-
-    client.put().uri("/api/v1.0/parents/{id}",Collections.singletonMap("id", parent.getId()))
+  public void updateTest() {
+    final Parennt parent = service.findByFullname_par("Sandrox").block();
+    final Parennt parentEditado = new Parennt("Sandro", "Masculino",
+        new Date(),"DNI", "20020711","22");
+    if (parent != null) {
+      client.put().uri("/api/v1.0/parents/{id}",Collections.singletonMap("id", parent.getId()))
         .contentType(MediaType.APPLICATION_JSON_UTF8)
         .accept(MediaType.APPLICATION_JSON_UTF8)
      .body(Mono.just(parentEditado), Parennt.class)
@@ -89,38 +114,55 @@ public void show() {
      .jsonPath("$.id").isNotEmpty()
      .jsonPath("$.fullname").isEqualTo("Sandro")
      .jsonPath("$.numberID").isEqualTo("20020711");
+    }
   }
 
+  /**
+ * Test findByFullname.
+ */
   @Test
-  public void buscarNombre() {
-    Parennt parent = service.obtenerPorName("David").block();
-    client.get()
-        .uri("/api/v1.0/students/nombre/{fullname}",
+  public void findByFullnameTest() {
+    final Parennt parent = service.findByFullname_par("David").block();
+    if (parent != null) {
+      client.get()
+        .uri("/api/v1.0/parents/nombre/{fullname}",
             Collections.singletonMap("fullname", parent.getFullname()))
         .exchange()
         .expectStatus().isOk()
         .expectBody().jsonPath("$.fullname").isEqualTo("David");
+    }
   }
-  
+
+  /**
+ * test findByNumberID.
+ */
   @Test
-  public void buscarDni() {
-    Parennt parent = service.obtenerPorName("Lucia").block();
-    client.get()
+  public void findByNumberIdTest() {
+    final Parennt parent = service.findByFullname_par("Lucia").block();
+    if (parent != null) {
+      client.get()
         .uri("/api/v1.0/parents/doc/{numberID}",
         Collections.singletonMap("numberID", parent.getNumberID()))
         .exchange()
         .expectStatus().isOk()
-        .expectBody().jsonPath("$.numberID").isEqualTo("20100711");
+        .expectBody().jsonPath("$.numberID").isEqualTo("20100711"); 
+    }
   }
-  
+
+  /**
+ * Test  delete.
+ */
   @Test
-  public void eliminarTest() {
-    Parennt parent = service.findByNumberID("00000001").block();
-    client.delete()
-        .uri("/api/v1.0/parents/{id}",Collections.singletonMap("id", parent.getId()))
-        .exchange()
-        .expectStatus().isOk()
-        .expectBody()
+  public void deleteTest() {
+    Parennt parent = service.findByNumberID("00000000").block();
+    if (parent != null) {
+      client.delete()
+      .uri("/api/v1.0/parents/{id}",Collections.singletonMap("id", parent.getId()))
+      .exchange()
+      .expectStatus().isOk()
+      .expectBody()
         .isEmpty();
+    }
+
   }
 }
